requirements.txt
pybit
pandas
numpy
tensorflow-cpu
scikit-learn
apscheduler
requests
nest_asyncio
pytz
